moved ner logic and functions to ner.py to not encumber basic.py
